Ran project through 2to3

+++ lib/vFense/plugins/monit/_db.py
@@ -115,7 +115,7 @@ def _save_data_point(agent=None, collection=None, data=None, conn=None):
     ):
         return None
 
-    timestamp = long(time.time())
+    timestamp = int(time.time())
 
     point = {}
     point[CollectionKeys.AgentId] = agent
